Create adxl345spi_with_chipselect.c
I have added the ability to select the "chipselect" for the SPI interface (CS_EN0 or CS_EN1) to make the usage of multiple ADXL345 possible. It should be noted that a simultenous usage is not possible since only one SPI interface is available. However, it is possible to access to ADXL345 in a series way. Maybe this will be useful for some users. In my use case I need two ADXL345, therefore this modification was necessary. 

Notes:
The new argument for selecting chipselect is "-c or --csel". The value to pass is an integer (0=CS_EN0, 1=CS_EN1).
If you wan to use the modified driver, please replace the original file with this one ! (dont forget to rename it in "adxl345spi.c")

opcodex64